tests/wpt/web-platform-tests/css/css-syntax-3/charset/page-windows-1251-css-at-charset-windows-1250-in-utf16.html

<!doctype html>
<title>CSS charset: page windows-1251, CSS @charset windows-1250 in utf-16</title>
<link rel=help href="https://drafts.csswg.org/css-syntax-3/#determine-the-fallback-encoding">
<meta charset=windows-1251>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel=stylesheet href="support/at-charset-windows-1250-in-utf16.css">
<div id=log></div>
<div id=&#xfffd;></div>
<div id=&#x0418;></div>
<div id=&#x010c;></div>
<script>
var elm_fffd = document.getElementById('\ufffd');
var elm_0418 = document.getElementById('\u0418');
var elm_010c = document.getElementById('\u010c');
var t = async_test();
onload = t.step_func(function(){
  assert_equals(getComputedStyle(elm_fffd, '').visibility, 'visible', 'selector U+FFFD matched (utf-8)');
  assert_equals(getComputedStyle(elm_0418, '').visibility, 'hidden', 'selector U+0418 did not match (windows-1251)');
  assert_equals(getComputedStyle(elm_010c, '').visibility, 'visible', 'selector U+010C matched (windows-1250)');
  this.done();
});
</script>
